[chore] modify websocket code framework
