+++ src/main/java/com/oneswap/controller/SystemController.java
@@ -1,6 +1,6 @@
 package com.oneswap.controller;
 
-import com.oneswap.websocket.InfraWeb3jClient;
+import com.oneswap.websocket.BlockchainWebSocketManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/1.0/system")
 public class SystemController {
 
-    private final InfraWeb3jClient infraWeb3jClient;
+    private final BlockchainWebSocketManager blockchainWebSocketManager;
 
     @GetMapping("/health")
     public ResponseEntity<?> healthCheck(){
@@ -23,7 +23,7 @@ public class SystemController {
 
     @GetMapping("/monitoring")
     public ResponseEntity<String> toggleMonitoring(@RequestParam boolean enable) {
-        infraWeb3jClient.setMonitoring(enable);
+        blockchainWebSocketManager.setMonitoring(enable);
         String message = enable ? "WebSocket monitoring started." : "WebSocket monitoring stopped.";
         return new ResponseEntity<>(message, HttpStatus.OK);
     }
